refactor: extract filtering logic into useProductFilters hook
