+++ src/components/ProductsList/index.tsx
@@ -3,19 +3,16 @@ import {
   useGetAllProductsQuery,
   useGetProductsByCategoryIdQuery,
 } from '../../store/api/apiSlice';
+import { useProductFilters } from '../../hooks/useProductFilters';
 import { categoriesMap } from '../../data/categories';
 import type { Product } from '@/types/data';
 import ProductCard from '../ProductCard';
 import ErrorMessage from '../ui/ErrorMessage';
 import styles from './ProductsList.module.css';
 import FilterPanelComponent from '../FilterPanel';
-import { useState } from 'react';
 import Spinner from '../ui/Spinner';
 
 export default function ProductList() {
-  const [priceRange, setPriceRange] = useState({ from: '', to: '' });
-  const [discountOnly, setDiscountOnly] = useState(false);
-  const [sortOrder, setSortOrder] = useState('default');
   const { categoryId } = useParams();
   const location = useLocation();
 
@@ -30,16 +27,6 @@ export default function ProductList() {
     isLoading: areAllProductsLoading,
     isError: areAllProductsError,
   } = useGetAllProductsQuery(undefined, { skip: !!categoryId });
-
-  const handlePriceChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setPriceRange((prev) => ({ ...prev, [e.target.name]: e.target.value }));
-  };
-  const handleDiscountChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setDiscountOnly(e.target.checked);
-  };
-  const handleSortChange = (value: string) => {
-    setSortOrder(value);
-  };
 
   let title: string = 'Products';
   let productsToRender: Product[] | undefined = [];
@@ -60,32 +47,15 @@ export default function ProductList() {
     productsToRender = allProducts;
   }
 
-  const filteredAndSortedProducts = productsToRender
-    ?.filter((product) => {
-      const productPrice = product.discont_price ?? product.price;
-      const from = parseFloat(priceRange.from);
-      const to = parseFloat(priceRange.to);
-      if (discountOnly && product.discont_price === null) return false;
-      if (!isNaN(from) && productPrice < from) return false;
-      if (!isNaN(to) && productPrice > to) return false;
-      return true;
-    })
-    .sort((a, b) => {
-      const priceA = a.discont_price ?? a.price;
-      const priceB = b.discont_price ?? b.price;
-      switch (sortOrder) {
-        case 'price_asc':
-          return priceA - priceB;
-        case 'price_desc':
-          return priceB - priceA;
-        case 'newest':
-          return (
-            new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
-          );
-        default:
-          return 0;
-      }
-    });
+  const {
+    priceRange,
+    discountOnly,
+    sortOrder,
+    handlePriceChange,
+    handleDiscountChange,
+    handleSortChange,
+    filteredAndSortedProducts,
+  } = useProductFilters(productsToRender);
 
   const isLoading = isCategoryLoading || areAllProductsLoading;
   const isError = isCategoryError || areAllProductsError;
